build!: migrate from CRA to Vite
[Create React App (CRA)](https://github.com/facebook/create-react-app)
has not been actively maintained for some time and relies on outdated
dependencies. For instance, it uses an older version of `globals` that
is incompatible with recent versions of ESLint. To address these issues,
I decided to migrate the project to a more modern build tool.

After evaluating [Vite](https://vite.dev/guide/) and
[Next.js](https://nextjs.org/), I chose Vite for its lightweight nature
and gentler learning curve.

This commit represents the initial step in the migration, ensuring the
web app is functional with Vite. Future commits will address:
- Migrating unit tests,
- Configuring formatting tools,
- Updating static file handling.

This change marks a significant improvement in the development
experience and modernization of the project.