refactor: replace Link and NavLink with button
Link (or NavLink) uses <a> tags under the hood. When mouse over, there will be a tiny (but noticable) tooltip in the left-bottom of your browser (maybe Chrome only) displaying the url of that <a> tag and it is not possible to hide it. It's quite anoying, as I am only doing client side routing, the url is of zero importance and will cover on the bottom buttons of my side bar.